Workaround for https://bugs.python.org/issue36094
Using mailproxy with Python 3.7+ results in the following error
when s.starttls() is called:
ValueError: server_hostname cannot be an empty string or start with a leading dot.

The workaround for this problem is to pass the target host and port to
the smtplib.SMTP() or smtplib.SMTP_SSL() constructors.